requirements.txt
grpcio==1.65.1
grpcio-tools==1.65.1
protobuf==5.27.2
setuptools==71.1.0
